Fixed picture file name in QC plot. Added
work-items: #48251, #48254
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C51952

+++ development/libraries/casaAPI/src/PrmCrustThinning.C
@@ -215,7 +215,7 @@ ErrorHandler::ReturnCode PrmCrustThinning::setInModel( mbapi::Model & caldModel 
          {
             throw ErrorHandler::Exception( caldModel.errorCode() ) << caldModel.errorMessage();
          }
-         if ( ErrorHandler::NoError != caldModel.setTableValue( s_crustIoTblName, i, s_crustIoTblOptimThicknessCol, 0 ) )
+         if ( ErrorHandler::NoError != caldModel.setTableValue( s_crustIoTblName, i, s_crustIoTblOptimThicknessCol, (long)0 ) )
          {
             throw ErrorHandler::Exception( caldModel.errorCode() ) << caldModel.errorMessage();
          }
